Ufs: factor out xattr code into xattr.rs
